Added: - LoadingscreenManager.cs - UIManager.cs - SceneManager.cs - CommandPrompt.cs
Removed: DataRemover.cs

Changed: some existing files.

+++ Assets/DataManagement/Scripts/Data/DataParser.cs
@@ -1,8 +1,4 @@
 ﻿using UnityEngine;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
-
 using System.IO;
 using System.Text;
 using System.Security.Cryptography;
@@ -15,8 +11,8 @@ namespace DataManagement
         {
             if (DataManager.Instance.encrypt)
             {
-                var t_inputbuffer = Encoding.Unicode.GetBytes(p_input);
-                var t_outputBuffer = DES.Create().CreateEncryptor(DataReferences.key, DataReferences.iv).TransformFinalBlock(t_inputbuffer, 0, t_inputbuffer.Length);
+                byte[] t_inputbuffer = Encoding.Unicode.GetBytes(p_input);
+                byte[] t_outputBuffer = DES.Create().CreateEncryptor(DataReferences.key, DataReferences.iv).TransformFinalBlock(t_inputbuffer, 0, t_inputbuffer.Length);
                 return System.Convert.ToBase64String(t_outputBuffer);
             }
             else return p_input;
@@ -24,23 +20,15 @@ namespace DataManagement
 
         public static ScriptableObject CreateAsset<T>(string p_name) where T : ScriptableObject
         {
-            var t_asset = ScriptableObject.CreateInstance<T>() as T;
-
-            #if UNITY_EDITOR
-            AssetDatabase.CreateAsset(t_asset, "Assets/Temp_Assets/" + p_name + ".asset");
-
-            AssetDatabase.SaveAssets();
-            AssetDatabase.Refresh();
-            EditorUtility.FocusProjectWindow();
-            Selection.activeObject = t_asset;
-            #endif
+            T t_asset = ScriptableObject.CreateInstance<T>() as T;
 
             return t_asset;
         }
 
         public static void SaveJSON(string p_name, string p_info)
         {
-            var t_path = Application.persistentDataPath + "/" + DataManager.Instance.DataReferences.ID + "/" + p_name + ".json";
+            if (DataManager.Instance == null) return;
+            string t_path = Application.persistentDataPath + "/" + DataManager.Instance.ID + "/" + SceneManager.Instance.DataReferences.ID + "/" + p_name + ".json";
             if (!File.Exists(t_path)) File.Delete(t_path);
 
             using (FileStream fs = new FileStream(t_path, FileMode.Create))
